Update the reporting for E2E contract test failures.
A common problem with the deployed contract E2E tests is when the
bytecode for a contract changes its hash and therefore its contract ID
changes.  These need to be updated in the source of the test depending
on the deployed contract.

This change is more verbose when there is a panic or revert for these
tests, to print the deployed contract IDs and the receipts which will
confirm a `ContractNotInInputs` error.