Enable LSP to work with multiple projects simultaneously (#3669)
Now that the `lazy_statics` have been removed from the type and
declaration engines, the language server can finally support providing
LSP capabilities for multiple simultaneous projects within the same
editor! 🥳

This allows the LSP tests to be run in parallel, and also allows for
unique projects to be opened in multiple windows within the editor.

Very excited to get this in. Should finally allow the language server to
work with non-trivial sway projects. cc @FuelLabs/application-dev

closes #3372 
solves the issues outlined in #3069