Remove pointers from the Sway IR. (#3663)
Remove pointers from IR.
    
The pointers in the IR are not really pointers, but are local variables.
We want to add proper pointers to the IR though and naming the local
variables as pointers adds ambiguity.
    
This change renames `get_ptr` to `get_local` and removes `ptr` and `mut`
from the serialised output.  `get_local` no longer performs casts so
`cast_ptr` has been added as a stopgap until we add a better solution.