Fixes issue with generic trait impls using the same method names.
When two generic trait implementation use the same method
implementation an error such as `Duplicate definition for method
"convert" for type "A"` would occur.

This fix does remove the error when a method with diferent type
parameters is added to another trait implementation with different type
parameters.
The fix also ensures that find_method_for_type returns a method whose
 parameters types matches the given arguments expressions return types.

Closes #3633.