Add the occurs check to prevent stack overflows with infinite type unification (#3686)
This PR adds an "occurs check" to prevent recursive type unification
causing stack overflow. The occurs check, in it's most simple form,
checks to see if generic variable _V_ exists in type _S_ to prevent a
case in which unifying _V_ and _S_ would create recursive
unification---i.e. we shouldn't unify generic type _V_ with
non-generic-type _S_ if _S_ contains _V_ anywhere in it's type
definition because that unification would recurse infinitely.

Here's a little bit more about the occurs check:
https://en.wikipedia.org/wiki/Occurs_check
And here is a bit about how it relates to type unification:
https://papl.cs.brown.edu/2016/Type_Inference.html

Closes #3317

Co-authored-by: emilyaherbert <emily.herbert@fuel.sh>