Adds docstring to core::never::Never
Copies the documentation from https://doc.rust-lang.org/std/primitive.never.html.
Copies the link above to the possible extend. Rust has more use cases to
the Never type than Sway currently has so the sections were not included in the
docstring as they are not applicable to the Sway Never type.