Merge `Type` and `Aggregate` into a copyable `Type`
This is the second step towards achieving #2819.

Combining Type and Aggregate into a single data structure
enables offset computations for complex indexing (i.e.,
a nesting of arrays and structs) which will be needed when
we introduce GEPs. The new function `get_indexed_type` does
this computation.

The core of this PR is in `irtype.rs`. Everything else is just
to accommodate that.