LSP: Save the `LexedProgram` AST to a temp file   (#3726)
This PR saves the `LexedProgram` AST to a temp file so it can be
displayed in VScode.

I have a PR open in the VScode plugin that adds a "Sway: Show Lexed AST"
command to the command palette. When that is executed, it loads the file
saved to temp memory that the server has saved.
https://github.com/FuelLabs/sway-vscode-plugin/pull/114

We currently have a command for showing the parsed and typed ASTs of the
currently open file, this just adds functionality for viewing the
initial lexing stage.