Pretty print `Vec<TyAstNode>`  (#3761)
We have a command in VScode called "Sway: Show typed AST". When
executed, it opens a 2nd window containing the typed AST for the
currently open sway file. This used to work but stopped behaving as
intended when the declaration engine was introduced.

Instead of showing the types, it just shows the declaration IDs of the
decl_engine. This PR just iterates over the `&[TyAstNode]` and looks up
the types in decl_engine and pretty prints them in order to return the
original functionality.