Improve the type engine API (#3758)
The goal of this PR is to reduce code bloat. So it does a number of
things, although there are no functional changes. Something to note is
that many of the code naming decisions (e.g. method names and such)
before this PR were influenced by the fact that the functions to access
the TE were global, so now having a more simple `TypeEngine::get` API
makes sense.

1. Rename `TypeEngine::insert_type` to `TypeEngine::insert`.
2. Rename `TypeEngine::look_up_type_id` to `TypeEngine::get`
3. Rename `TypeEngine::resolve_type` to `TypeEngine::resolve`
4. Rename files
5. Remove the stale `ResolvedType` type

Co-authored-by: emilyaherbert <emily.herbert@fuel.sh>