Introduce Configuration-time Constants (#3536)
Closes https://github.com/FuelLabs/sway/issues/1498

This PR adds configuration-time constants as described in Ref
https://github.com/FuelLabs/sway-rfcs/pull/19. There is still some work
to be done, namely:
- https://github.com/FuelLabs/sway/issues/3680
- https://github.com/FuelLabs/sway/issues/3679
- https://github.com/FuelLabs/sway/issues/3678
- https://github.com/FuelLabs/sway/issues/3306 (expanding that issue to
consider all types of constants)
as well as some work in the SDKs and in `forc` to override these
constants at deploy-time.

The design of this feature is relatively simply:

### Key Points
- Configuration-time constants are declared using a `configurable` block
which is quite similar to a `storage` block. It also requires that each
config constant is initialized (just like storage variables).
- Configuration-time constants show up in the JSON ABI as described in
the [ABI
spec](https://fuellabs.github.io/fuel-specs/master/protocol/abi/json_abi_format.html)
- There is a 1:1 mapping between each configuration-time constant and a
particular location in the data section. The amount of bytes consumed in
the data section is dependent on the type of the constant.
- The optimizer guarantees that configuration-time constants are not
optimized in any way.
- The IR represents configuration-time constants are global scope (see
IR tests)
- Configuration-time constants are not allowed in libraries for the time
being.

### Parser
Converts a `configurable` block into a list of constant declaration with
a new field that indicates whether the constant is configurable or not.
In other words, the concept of a configurable block does not exist after
`convert_parse_tree`.

### Frontend
The type checker is virtually unchanged.

### IR Gen and IR
Simply special casing global constants that have `is_configurable =
true` and converting those to a special type of `Value`s called
`Configurable`. This makes sure that the optimizer knows not to touch
this type of values in any way.

I also had to add a new metadata to keep track of the name of each
config. This is important to know what data value corresponds to what
config const in codegen because that information is useful for the JSON
object.

### Codegen
Mostly the same as regular constants except that we need to be careful
about merging data values that are the same. We continue to merge values
for regular consts but we make sure that configuration consts are
uniqued and get their own unshared spot.

### JSON ABI
The configuration consts and their types are added to the JSON object in
the frontend as usual. We later modify the object and add the actual
bytecode offsets as computed in codegen.

### Raw Tasks:
- [x] Parser
- [x] Frontend
- [ ] (deferred) Disallow multiple `configurable` blocks: will do this
separately in https://github.com/FuelLabs/sway/issues/3678
- [x] new IR `Value` type: `Configurable`
  - [x] Printing
  - [x] Parsing
  - [x] New metadata `config_name` 
- [x] IR Gen, especially on `const_eval`
  - [x] Unit tests 
- [x] ASM gen
- [ ] (deferred) Make sure unused configurable variables are still
included in the data section: See
https://github.com/FuelLabs/sway/issues/3680
  - [x] Unit tests
- [x] Handling of data section and returning a map: config name ->
bytecode offset
- [x] JSON ABI generation + `forc-pkg` changes
- [x] E2E tests
- [ ] (deferred) Formatter. See:
https://github.com/FuelLabs/sway/issues/3679