Add `ops::*` to the prelude in `core` (#3892)
This PR fixes the bug described in #3891 by adding `ops` to the prelude
in `core`. This change is congruent with Rust's approach as well.

This is a _breaking change_---if any users have any traits manually
implemented for which core already has an existing implementation, this
will now create an error. A rightful one, IMO, but a breaking change
nonetheless.

A few of the tests were written in a way that is not compatible with
this change---so this PR rewrites these. The only one of note is that a
old `should_pass/language/trait_override_bug` test was refactored to a
new `should_fail/redefine_method_from_core` test.

This PR also refactors the trait finding algorithm in
`check_if_trait_constraints_are_satisfied_for_type` to reduce it from
O(n^2) to O(n).

Closes #3891

Co-authored-by: emilyaherbert <emily.herbert@fuel.sh>