Improve performance of the register allocator (4.5X) (#3909)
- Avoid the use of `remove_node` by invalidating nodes in the
interference graph instead of actually removing them. For that, I now
use an `Option` for the node weight; when the weight is `None`, the node
is invalid and is not considered anywhere in the algorithm.
- In the register coalescing algorithm, when coalescing two nodes `ix1`
and `ix2`, _merge_ `ix2` into `ix1` instead of merging them both into a
separate third node. This involves copying the edges connected to `ix2`
into `ix1` and the invalidating `ix2` by setting its weight to `None`.
- Avoid cloning as much as possible by using hash maps of references.

I double checked that the bytecode size is unchanged for several small
and large designs so I know that the allocator is not less efficient in
that sense.

For https://github.com/sway-libs/concentrated-liquidity, the
improvements are as follows for the IR -> final bytecode step:
- Before this change: 192.49 seconds
- After this change: 42.72 seconds

**for a 4.5X improvement.**

Most of the remaining time (really like 99% of it) is spent in computing
`live_out` (i.e. in
[here](https://github.com/FuelLabs/sway/blob/master/sway-core/src/asm_generation/fuel/register_allocator.rs#L130)).
I will be looking at that separately.