feat(fmt): introduce formatting comments together with src (#3854)
closes #3853
closes #3574


## Overview

Currently, comment formatting is done post-formatting of the source
code, which can make it difficult to reason about formatting comments in
certain scenarios. There have been a few cases
(https://github.com/FuelLabs/sway/issues/3574,
https://github.com/FuelLabs/sway/issues/2888,
https://github.com/FuelLabs/sway/issues/2649) where comments are often
wrongly indented.

In #3789, there is extensive discussion on an alternative way of
handling comment formatting. It seems like we're at a similar problem
that rustfmt originally encountered since comments are not part of the
AST.

## Approach

Instead of formatting comments _after_ formatting the source code as a
whole w/o comments, we now format comments locally within each `Format`
implementation of each AST node. Few advantages of doing this:

1) We now have the context of the formatter at the point of which the
comments are being added, e.g. we can use `formatter.shape.indent` to
indent comments, especially in odd locations.

2) Easier handling of `LineStyle::Multiline` <> `LineStyle::Inline`
conversions, when comments are involved. You can see this from how we
can check if comments are present and we can decide which `LineStyle` to
use in the `}` and `else` formatting of comments.

To make sure that we don't do too much extra work, we use
`comment_map_from_src` to init a `CommentMap` at the point where
formatting begins instead of during `handle_comments` - we then reuse
this `CommentMap` for the entire pass of the formatting. This is
temporary and the end goal is to deprecate comment insertion
post-formatting the source code, and to do it all in one pass.

To ensure backward compatibility, we _still_ call `handle_comments` for
now until we refactor everything to use the new method. Thankfully we
have extensive tests that will prevent regression while this refactor is
happening.

This will be a gradual refactor and the above 2 issues were selected
since they were difficult to fix with the old model of comment
insertion, but much easier with this new model.