Fix handling of empty documents (#3930)
Closes #3923 

The name here is slightly misleading. This does clean up how we handle
it to a degree, but mostly due to a conflict of what is considered
documentable. We don't have html rendering for `ImplTrait`s just yet,
and because they were considered documentable but not handled we were
getting a panic.

While working this out, I found another bug: #3929

Co-authored-by: Joshua Batty <joshpbatty@gmail.com>