refactor(fmt): move tests (#3959)
groundwork for #3597

Move tests from `formatter/mod.rs` to a dedicated tests directory. There
are no logic changes aside from the move.

This is to prepare for an adjacent `macros.rs` file that contains the
code for the above.

Plus `formatter/mod.rs` has reached over 1.5k LoC, so it's not a bad
idea to slim it down, and we get to trim off 1 level of indentation too.