refactor(fmt): cleanup testsuite using helper fn and macro (#3960)
closes #3597

It's not entirely accurate that this addresses the issue, since the
usage of a macro here (credits to
[ripgrep](https://github.com/BurntSushi/ripgrep/blob/fe97c0a152cabc1bc07ec36b4b1e27cd230c3014/tests/macros.rs#L18))
is only to improve diff printing, and all the setups are abstracted into
functions, which are reused throughout the tests. I'm not sure that we
need macros here for those purposes, so I did not use them, but the
functions themselves accomplish the goal of reducing boilerplate anyway,
so I am marking it as closing that issue ;)

I've also renamed the fn names. Some of their names are inconsistent -
some start with 'test_...', some start with the subject being tested
e.g. 'fn struct_...' or 'fn comments_...' since the tests are already
labelled with `#[test]` I removed the 'test_' prefix from all functions.