Rely on the `IfExp` ast node return type in IR gen (#3978)
## Description
Closes #3970 

In IR gen, rely on the type of the AST node `IfExpr` when deciding on
the type of the block argument. The type checking phase is responsible
for deciding what that type is and not IR gen. This was required to fix
the issue linked above.

## Checklist

- [x] I have linked to any relevant issues.
- [x] I have commented my code, particularly in hard-to-understand
areas.
- [x] I have added tests that prove my fix is effective or that my
feature works.
- [x] I have done my best to ensure that my PR adheres to [the Fuel Labs
Code Review
Standards](https://github.com/FuelLabs/rfcs/blob/master/text/code-standards/external-contributors.md).
- [x] I have requested a review from the relevant team or maintainers.

---------

Co-authored-by: João Matos <joao@tritao.eu>