Fixes regression caused by enum and struct callpaths.
The issue is that the module in which a struct or enum is can be
imported with different callpaths.

For instance we could have a struct with callpath
`my_script::data_structures::SomeStruct` in a module and in another
module as `my_script::eq_impls::data_structures::SomeStruct`. This makes
callpaths unreliable to compare declarations.

Closes #3998

Reopens #418

Disables tests multiple_enums_with_the_same_name and
multiple_structs_with_the_same_name.