Add an extra mdbook build step in attempt to get readable output
The mdbook forc documenter step currently dumps output to a file, so we
can't see what actually went wrong, only that something did.

This adds a clone of this step to see if we can see the output.

Also removes unnecessary ERROR log check from doc reference, as it does
not use the mdbook-forc-documenter plugin at all.