Introduce a `StorageBytes` type in the standard library (#4001)
## Description

The following allows for the storage of a collection of tightly packed
bytes with the use of the `Bytes` type. This will provide significant
space savings over a `StorageVec<u8>` and will also unblock
`StorageString` in the Sway-libs repo.
Please note however that the bytes are still padded to the size of a
quadword in storage. There will be at most 31 padded bytes for any
`StorageBytes` type.

I've initially opted to omit adding functions such as `push()`. This is
to encourage any byte manipulation to be done in the heap using the
`Bytes` type and prevent successive storage read/writes calls in the
same transaction.

Closes #3910 

## Checklist

- [x] I have linked to any relevant issues.
- [x] I have commented my code, particularly in hard-to-understand
areas.
- [x] I have updated the documentation where relevant (API docs, the
reference, and the Sway book).
- [x] I have added tests that prove my fix is effective or that my
feature works.
- [x] I have added (or requested a maintainer to add) the necessary
`Breaking*` or `New Feature` labels where relevant.
- [x] I have done my best to ensure that my PR adheres to [the Fuel Labs
Code Review
Standards](https://github.com/FuelLabs/rfcs/blob/master/text/code-standards/external-contributors.md).
- [x] I have requested a review from the relevant team or maintainers.

---------

Co-authored-by: bitzoic <cameron.carstens@fuel.sh>