Add a new `name` field on `DeclId` and remove some unnecessary lookups in the `DeclEngine` (#4028)
## Description

This PR is a subset of #3744. It adds a `name` field to `DeclId` which
contains the name of the declaration. This is helpful in that it allows
us to bypass lookups in the `DeclEngine` for trivial name info and is
also a subset of #3744.

This PR also removes/refactors some of these unnecessary lookups in the
`DeclEngine` in light of the new `name` field.

## Checklist

- [x] I have linked to any relevant issues.
- [ ] I have commented my code, particularly in hard-to-understand
areas.
- [x] I have updated the documentation where relevant (API docs, the
reference, and the Sway book).
- [ ] I have added tests that prove my fix is effective or that my
feature works.
- [x] I have added (or requested a maintainer to add) the necessary
`Breaking*` or `New Feature` labels where relevant.
- [x] I have done my best to ensure that my PR adheres to [the Fuel Labs
Code Review
Standards](https://github.com/FuelLabs/rfcs/blob/master/text/code-standards/external-contributors.md).
- [x] I have requested a review from the relevant team or maintainers.

---------

Co-authored-by: emilyaherbert <emily.herbert@fuel.sh>