refactor: Abstract out `source`-related items from `pkg` in `forc-pkg`
In anticipation of the upcoming ipfs support, this PR refactors `forc-
pkg` to more clearly separate out the "source" abstraction from the rest
of the pkg-handling logic.

We do so by creating a common interface for sources in the form of
traits for pinning and fetching. This should allow for adding in the
upcoming ipfs implementation (or future alternative source kind) in a
manner that doesn't further complicate the rest of package handling.

There is still room for further refactor of `pkg.rs`, but this should do
enough to clear the way for ipfs support and close #3927.