DCA now detects dead code when all branches return explicitly.
Fixing `If` straightforward. But that didn't automatically the fix
`match` scenario (see test case). Turns out that `match` always
has a covering "ImplicitReturn" expression, and with implicit returns
in DCA, we add an edge from the beginning to the end, which I couldn't
figure out why. Just removing that got it working. I request the
reviewers to validate this part of the change carefully.

Closes #2313