feat(forc-pkg): Improve workspace build summary
Previously when building a large workspace, the build would often end
followed by a bunch of lines saying "Bytecode size is ..." or "Bytecode
hash is ...", without any indication of which workspace member the
information was associated with.

This changes the behaviour to first output a "Finished" line, followed
by build output summary info clearly separated by the workspace member.