Generic refactor of `DeclEngine`
Refactor the `DeclEngine` to use generic `DeclId`s that carry the type
of the declaration they point to with them. This allows us to avoid a
*lot* of unnecessary error handling and passing around of spans.

`DeclId` types mostly to not mix except in the case of Functions/TraitFns and Traits/Abis.
In these cases we use `FunctionalDeclId` and `InterfaceDeclId` enums to mix them in a
way that can still be discriminated by the user.