refactor(forc-pkg): Attempt to simplify compile fns with dedicated types
This introduces the `CompiledPackage` and `CompiledContractDependency`
types in order to more accurately model the inputs and outputs of
the `compile` function.

We also move the `name` and `target` fields from `BuiltPackage` into
the `PackageDescriptor`, construct the descriptor earlier in the build
process, and use the descriptor to consolidate more of the `compile`
function arguments.

I started on refactoring some of the ABI generation related code, but
realised that this might involve a more dedicated effort and opened