LSP: Rewrite `LexedProgram` and `ParseProgram` traversal to use the Parse trait. (#4220)
## Description
This work was started in #3772. I've decided to split that up into 2
PR's, this one implements a new `Parse` trait for the lexed and parsed
ast's. The second one coming up with do the same but for the typed ast.

The diff is quite big but it's just essentially moving the logic that
was there into the trait impl's.

Partially addresses #3799