Add lock file for sdk-harness test workspace, fix fuel-core update CI issues
The lock file pins the fuel-core crates to 0.17.3, there appears to
be some breakage in 0.17.4:
https://github.com/FuelLabs/sway/actions/runs/4371254731/jobs/7646959456

I tried adding the sdk-harness test crates to the top-level workspace,
but this caused loads more unrelated errors which I'm not sure how to
address just yet. This PR should at least get CI working again for now.